Move folder from integration-test to src/test
