+++ .gitignore
@@ -50,5 +50,5 @@ src/rttmap.csv
 log/
 *.history
 
-integration-test/profile.cov
-integration-test/meta_db/dat/*
+src/test/profile.cov
+src/test/meta_db/dat/*
