+++ src/test/cli-scenario/cli_misc_json_test.go
