+++ src/test/cli-scenario/cli_misc_yaml_test.go
