+++ src/test/cli-scenario/setup.go
