+++ src/test/go-scenario/go_misc_test.go
