+++ src/test/go-scenario/method_call.go
