+++ src/test/rest-go-equal-scenario/2_rest_go_equal_test.go
