+++ src/test/test.env
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-export CBTUMBLEBUG_ROOT=$HOME/go/src/github.com/cloud-barista/poc-cicd-tumblebug/integration-test
+export CBTUMBLEBUG_ROOT=$HOME/go/src/github.com/cloud-barista/poc-cicd-tumblebug/src/test
 export CBSTORE_ROOT=$CBTUMBLEBUG_ROOT
 export CBLOG_ROOT=$CBTUMBLEBUG_ROOT
 export SPIDER_CALL_METHOD=REST
